Improve USER situation:
We moved the USER app to a ONBUILD rule, thus allowing the first hook,
after-init-hooks, to still run as root.

This allows the hook to install system packages.

After the hook is executed, we make sure the entire /app has proper
ownership for the follow-up hooks. Please note that files COPY'ed over
will still be owned by root.

Signed-off-by: Pedro Melo <melo@simplicidade.org>