Keep local/ in a separate place
After we install cpanfile deps, we move the local/ to /app/deps and
we let the entrypoint script move if to the correct place.

This allows us to run our container but mount a local app folder and
still use the container dependencies.

If a local/ exists in the local app folder, it is moved to
a `local.<timestamp>`.